Functia lui Ackley/requirements.txt
matplotlib==3.9.3
numpy==2.1.3
python-dotenv==1.0.1